image-classifier/requirements.txt
tensorflow==2.13.0
keras==2.13.1
flask==2.3.3
pillow==10.0.1
numpy==1.24.3
matplotlib==3.7.2
scikit-learn==1.3.0
jupyter==1.0.0
opencv-python==4.8.1.78
